+++ lib/karafka/routing/features/active_job/proxy.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Karafka
+  module Routing
+    module Features
+      class ActiveJob < Base
+        # Routing proxy extensions for ActiveJob
+        module Proxy
+          include Builder
+        end
+      end
+    end
+  end
+end