+++ spec/integrations/pro/consumption/strategies/dlq/ftr_lrj_mom/non_recoverable_moving_forward_marking_spec.rb
@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
 class Listener
-  def on_error_occurred(event)
-    DT[:errors] << event
+  def on_error_occurred(_)
+    DT[:errors] << true
   end
 end
 
 Karafka.monitor.subscribe(Listener.new)
 
 setup_karafka(allow_errors: true) do |config|
-  config.max_messages = 10
+  config.max_messages = 1
   config.kafka[:'max.poll.interval.ms'] = 10_000
   config.kafka[:'session.timeout.ms'] = 10_000
 end